move job map tests to viz tests
