+++ tests/test_job.py
@@ -1,11 +1,8 @@
 import os
 from pathlib import Path
 import tempfile
-import tarfile
-from shutil import copyfile
 
 import pytest
-from folium import Map
 
 # pylint: disable=unused-import
 from .context import Job, JobTask
@@ -232,35 +229,3 @@ def test_job_download_result_live_2gb_big_exceeding_2min_gcs_treshold(auth_live)
         for file in out_files:
             assert Path(file).exists()
         assert len(out_files) == 490
-
-
-def test_map_results(job_mock):
-    fp_tgz = Path(__file__).resolve().parent / "mock_data/result_tif.tgz"
-    with tarfile.open(fp_tgz) as tar:
-        tar.extractall(fp_tgz.parent)
-    fp_tif = (
-        fp_tgz.parent
-        / "output/7e17f023-a8e3-43bd-aaac-5bbef749c7f4/7e17f023-a8e3-43bd-aaac-5bbef749c7f4_0-0.tif"
-    )
-    fp_data_json = fp_tgz.parent / "output/data.json"
-
-    job_mock.results = [str(fp_tif), str(fp_data_json)]
-    map_object = job_mock.map_results()
-    assert isinstance(map_object, Map)
-
-
-def test_map_results_additional_geojson(job_mock):
-    fp_tgz = Path(__file__).resolve().parent / "mock_data/result_tif.tgz"
-    with tarfile.open(fp_tgz) as tar:
-        tar.extractall(fp_tgz.parent)
-    fp_tif = (
-        fp_tgz.parent
-        / "output/7e17f023-a8e3-43bd-aaac-5bbef749c7f4/7e17f023-a8e3-43bd-aaac-5bbef749c7f4_0-0.tif"
-    )
-    fp_data_json = fp_tgz.parent / "output/data.json"
-    fp_data_geojson = fp_tgz.parent / "output/additional_vector_file.geojson"
-    copyfile(fp_data_json, fp_data_geojson)
-
-    job_mock.results = [str(fp_tif), str(fp_data_json), str(fp_data_geojson)]
-    map_object = job_mock.map_results()
-    assert isinstance(map_object, Map)
